libmpathpersist: split public and internal API
The libmpathpersist library exports symbols that are not part
of the public API. Move the respective code into a separate
source file.
